Support new access_token in query string (#13)
* Support new access_token in the

Access_token needs to be supported in order for the websocket connection
to success from a web browser. The API within web browsers cannot set
custom headers. So, this is the workaround provided for now until
the API for web browsers is updated somehow to allow for setting
bearer tokens in the auth headers.

See: https://tools.ietf.org/html/rfc6750#section-2.3
     https://github.com/cashweb/cashweb-rs/pull/3

Fixes #12

* Little refactor

* Bump dependencies

Co-authored-by: Harry Barber <harrybarber@protonmail.com>