Refactor Bitcoin Client API (#8)
Currently, the API exposes the inner errors which are all template
bound. This makes it very hard to properly wrap and bubble errors from
other parts of the codebase without implementing larges amounts of
traits unnecessarily. This also makes it so that the implementation is
leaky w/r/t libraries it uses.

This commit refactors the Bitcoin Client library to not require
templates, and it also no longer leaks error types from underlying
libraries.