Fix voting api to not discard older burn transactions
Currently, there is a bug where all old burn transactions are replaced.
However, this is not the desired behavior. This commit properly
consolidates burn transactions that were recorded with the new burn
transactions before saving the post back to the database. There is one
remaining issue in that this should occur within a RocksDB transaction.